+++ imgui.ini
@@ -9,7 +9,7 @@ Size=476,528
 Collapsed=1
 
 [Window][Editor]
-Pos=10,32
+Pos=10,34
 Size=316,475
 Collapsed=0
