+++ src/graphics/ui.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "window.h"
+#include "window.hpp"
 
 namespace sand {
